+++ hoomi/jobs/models/__init__.py
@@ -1,2 +1,3 @@
 from .job import Job
 from .photo_job_history import PhotoJobHistory
+from .experience import Experience
