+++ hoomi/jobs/models/abstract_job_history.py
@@ -9,14 +9,6 @@ class AbstractJobHistory(models.Model):
 
     theme = models.IntegerField()
 
-    content = models.CharField(max_length=100,)
-
-    hash_id = models.CharField(
-        max_length=5,
-        blank=True,
-        null=True,
-    )
-
     created_at = models.DateTimeField(auto_now_add=True)
 
     updated_at = models.DateTimeField(auto_now=True)
